// src/app/starwars/interfaces/interfaces.ts
export interface Starship {
    count:    number;
    next:     string;
    previous: null;
    results:  Result[];
}

export interface Result {
    
    name:                   string;
    model:                  string;
    manufacturer:           string;
    cost_in_credits:        string;
    length:                 string;
    max_atmosphering_speed: string;
    crew:                   string;
    passengers:             string;
    cargo_capacity:         string;
    consumables:            string;
    hyperdrive_rating:      string;
    MGLT:                   string;
    starship_class:         string;
    pilots:                 string[];
    films:                  string[];
    created:                Date;
    edited:                 Date;
    url:                    string;
}
export interface User {
    id:          number;
    name:   string;
    email:        string;
    password:     string;

  }